requirements.txt
streamlit>=1.41.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.5.0
pdfkit>=1.0.0
markdown>=3.4.3
mdformat>=0.7.16
streamlit-pdf-viewer>=0.0.20
pdfplumber>=0.11.5
